feat(timeline): extend existing locales
Extend existing locales instead of replacing them. Fix #55.

This also extends all locales to use the default locale for any
missing key. So if a new localization option is added in the future,
every locale which does not yet have the new option translated will use
the default english one.