Catch when navigator is undefined - fixes #231
`if (!navigator)` throws an error when `navigator` has never been declared before.

Another option is to to check `typeof navigator === 'undefined'`. I chose to do a try...catch because I also want it to return `en` for any falsy value and not just `undefined` (if `navigator` were declared, but `null` for example).