feat(build): remove default export from standalone
BREAKING CHANGE: The default export is no longer provided.

This doesn't do much anyway.

In ESM it's possible to do:
  `import * as visTimeline from "vis-timeline/standalone";`

In UMD it just adds a property named `default` which contains the same
stuff as the export itself does already. I highly doubt anyone is
actually using this in UMD.