feat(build): add peer, standalone and esnext builds (#269)
* feat(build): add peer and standalone builds

* chore(readme): fix out of place some references to Network

* ci: update gh-pages copied paths

* ci: persist the new dirs

* docs(examples): use standalone build

* feat(build): add ESNext build

This is a pretty small investment on our side but can help a lot to the
users who want to optimize their bundle size. Well, when it will be
deployed in the other projects anyway.

* fix(build): do not remove locale from legacy

* feat(build): remove default export from standalone

BREAKING CHANGE: The default export is no longer provided.

This doesn't do much anyway.

In ESM it's possible to do:
  `import * as visTimeline from "vis-timeline/standalone";`

In UMD it just adds a property named `default` which contains the same
stuff as the export itself does already. I highly doubt anyone is
actually using this in UMD.

* chore(build): add dev dep import safeguard

* chore(readme): some minor edits and additions