fix: update lastStart on every graph update
Normally dragging a line graph (or calling setRange repeatedly to sync
the range with something external) causes only partial updates (the SVG
is moved as a whole, not modified). However, if at the same time full
updates happen the graph can become moved twice: once because the items
have been moved and the second time due to the SVG offset.

To avoid this problem this commit updates the lastStart reference
position not only when the drag ends, but also on any other full graph
rebuild.