feat(data): allow custom DataSet/View implementations (#729)
This tests that the methods and properties are implemented instead of
checking the prototype. Thanks to this it is now possible to use a
different DataSet or DataView without extending the Vis one. This also
removes some interoperability issues of the standalone build when used
toghether with other Vis family libraries.