Keep visibleItems array sorted
performStacking performs better based on how close the ordering of the input array is to ascending start times. The input to performStacking is derived from the Group.visibleItems array.

This array is sometimes sorted by start time (for example, when items are added to it via the _traceVisible method and this method was given orderedItems.byStart as an argument). But other times (such as when _traceVisible is given orderedItems.byEnd, or when the _checkIfVisible method is called) items are not added to the array in any useful order.

This change keeps the visibleItems array sorted by start time. In a dataset of 6,000 items with no custom ordering, I measured a 1-4% reduction in stacking time from this change.