Update _traceVisible to preserve item order
The return value of _traceVisible is frequently passed to performStacking. performStacking performs better based on how close the ordering of the input array is to ascending start times.

Prior to this change, half of _traceVisible's output would be ordered consistently with the input order, but half would be reversed. After the change, the output order is consistent with the input order.

Testing in Firefox on a 6,000 item dataset with no custom order function showed that this change reduced stacking time by a modest ~5%. When a custom order function was provided, no performance improvement was measured.