Fix performance bug in stacking algorithm
This fixes several bugs and oversights in the stacking algorithm. None of these affected accuracy but they did affect performance.

- Since the horizontalOverlapEndIndex value is precise (unlike horizontalOverlapStartIndex, which is a conservative approximation) we don't need to check each item's start in the `filterBetween` call.

- previousEnd was never set, which meant that the horizontalOverlapEndIndex would never be moved backwards. (This was not noticed earlier because the previous oversight hid the resulting issues.)

- There was a capitalisation typo (`horizontalOVerlapEndIndex`) but it was in code rendered unreachable by the previous bug.

- There was an off-by-one error where inserting an item into an array made the tracked indexes incorrect. The impact of this was also not noticed due to the previous bugs.

Overall this commit on average reduces stacking time by 4% when a custom order function is in use (tested with a set of 6,000 items). There is no change when a custom order function is not provided.