feat(timeline): improve stack function performance (#1847)
* Update _traceVisible to preserve item order

The return value of _traceVisible is frequently passed to performStacking. performStacking performs better based on how close the ordering of the input array is to ascending start times.

Prior to this change, half of _traceVisible's output would be ordered consistently with the input order, but half would be reversed. After the change, the output order is consistent with the input order.

Testing in Firefox on a 6,000 item dataset with no custom order function showed that this change reduced stacking time by a modest ~5%. When a custom order function was provided, no performance improvement was measured.

* Remove expensive uses of Array.slice

When I wrote #1281 I was under the impression that calling `Array.slice` and then using the resulting array in a read-only capacity would be optimised by JS engines to use an offset/crop into the same memory as the original array, similar to how Node's Buffer.subarray works.

I have since discovered that this is not the case: in both Chrome & Firefox,`Array.slice` always carries the cost of shallow-copying the array. Avoiding the use of `slice` significantly improves stacking performance with large amounts of items.

(On a dataset of 6,000 items, this reduces stacking time by 5-20%!)

* Keep visibleItems array sorted

performStacking performs better based on how close the ordering of the input array is to ascending start times. The input to performStacking is derived from the Group.visibleItems array.

This array is sometimes sorted by start time (for example, when items are added to it via the _traceVisible method and this method was given orderedItems.byStart as an argument). But other times (such as when _traceVisible is given orderedItems.byEnd, or when the _checkIfVisible method is called) items are not added to the array in any useful order.

This change keeps the visibleItems array sorted by start time. In a dataset of 6,000 items with no custom ordering, I measured a 1-4% reduction in stacking time from this change.

* Fix performance bug in stacking algorithm

This fixes several bugs and oversights in the stacking algorithm. None of these affected accuracy but they did affect performance.

- Since the horizontalOverlapEndIndex value is precise (unlike horizontalOverlapStartIndex, which is a conservative approximation) we don't need to check each item's start in the `filterBetween` call.

- previousEnd was never set, which meant that the horizontalOverlapEndIndex would never be moved backwards. (This was not noticed earlier because the previous oversight hid the resulting issues.)

- There was a capitalisation typo (`horizontalOVerlapEndIndex`) but it was in code rendered unreachable by the previous bug.

- There was an off-by-one error where inserting an item into an array made the tracked indexes incorrect. The impact of this was also not noticed due to the previous bugs.

Overall this commit on average reduces stacking time by 4% when a custom order function is in use (tested with a set of 6,000 items). There is no change when a custom order function is not provided.

* Update uses of var to let/const

* Fix typo introduced in conflict resolution