chore(index): add trailing slashes to the link in index
This is used with a local server during development so that you can just
click your way to examples and don't have to type in the path. On our
site we use the trailing slash and the example index pages don't work
without it since they use ../ to load images and link to examples.